[Fission] Attempt to add support for multiple configmaps and volume mounts.
[git-p4: depot-paths = "//depot/service/java/projects/DistributedFileHandling/ContentProcessing/modules/kubeclient/": change = 538793]

+++ fabric8/main/src/main/java/com/theplatform/dfh/cp/modules/kube/fabric8/client/PodPushClientImpl.java
@@ -184,6 +184,11 @@ public class PodPushClientImpl implements PodPushClient
         try
         {
             PodResource<Pod, DoneablePod> podResource = kubernetesHttpClients.getRequestClient().getPodResource(kubeConfig.getNameSpace(), podName);
+            if(podResource.get() == null)
+            {
+                logger.warn("Attempted to delete missing pod [{}]", podName);
+                return true;
+            }
             logPodSpecs(podResource.get(), DELETE_POD_TEMPLATE);
             podResource.delete();
             return true;
